attempt to use kooparse/zgltf binding lib
this is an unfinished attempt to parse res/testcube.gltf using zgltf.

i've been able to add a zgltf module and parse res/testcube.gltf with it
but i'm not sure how to parse the vertices.  maybe you can figure it out
if you think this is preferrable to using cgltf directly.